apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-pptx==0.6.23
supabase==2.3.4
openai==1.10.0
openpyxl==3.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
httpx==0.26.0

